feat: move face off image into file and shuffle positions once
